Отчет Allure

+++ src/main/java/requests/steps/UserSteps.java
@@ -54,7 +54,7 @@ public class UserSteps {
         return createdAccounts;
     }
 
-    public static void successDepositToUserAccount(CreateUserRequest userRequest, Integer userId, double balance) {
+    public static void successDepositToUserAccount(CreateUserRequest userRequest, Long userId, double balance) {
         new CrudRequester(authSpec(userRequest),
                 Endpoint.DEPOSIT,
                 ResponseSpecs.requestReturnsOK())
@@ -65,7 +65,7 @@ public class UserSteps {
     }
 
     public static void failDepositToUserAccount(CreateUserRequest userRequest,
-                                                Integer accountIdUser,
+                                                Long accountIdUser,
                                                 double balance,
                                                 ResponseSpecification responseSpecs) {
         new CrudRequester(authSpec(userRequest),
@@ -142,8 +142,8 @@ public class UserSteps {
                 .build();
     }
 
-    private static TransferUserDepositRequest buildTransferUserDeposit(Integer accountIdSenderUser,
-                                                                       Integer accountIdReceiverUser,
+    private static TransferUserDepositRequest buildTransferUserDeposit(Long accountIdSenderUser,
+                                                                       Long accountIdReceiverUser,
                                                                        Double amount) {
         return TransferUserDepositRequest.builder()
                 .senderAccountId(accountIdSenderUser)
